Final Commit!

+++ src/InputForm.js
@@ -1,16 +1,19 @@
 import React, { useState, useEffect } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import ServerResponseModal from './ServerResponseModal';
+import BorderExample from './GrowExample'; 
+
 
 const InputForm = () => {
   const [uploadCount, setUploadCount] = useState(3);
   const [serverResponse, setServerResponse] = useState(null);
   const [showModal, setShowModal] = useState(false);
   const [isFileSelected, setIsFileSelected] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
 
   useEffect(() => {
     let userUUID = localStorage.getItem('userUUID');
-
+  
     if (!userUUID) {
       userUUID = uuidv4();
       localStorage.setItem('userUUID', userUUID);
@@ -21,9 +24,12 @@ const InputForm = () => {
         setUploadCount(parseInt(savedUploadCount, 10));
       }
     }
-
+  
     if (parseInt(uploadCount, 10) === 0) {
-      document.getElementById('submitBtn').disabled = true;
+      const submitBtn = document.getElementById('submitBtn');
+      if (submitBtn) {
+        submitBtn.disabled = true;
+      }
     }
   }, [uploadCount]);
 
@@ -35,17 +41,20 @@ const InputForm = () => {
   const handleShowModal = () => setShowModal(true);
 
   const handleUpload = async () => {
+    // Check if a file is selected
+    if (!isFileSelected) {
+      alert('Please select a file before submitting.');
+      return;
+    }
+
     const userUUID = localStorage.getItem('userUUID');
     const newUploadCount = uploadCount - 1;
 
     setUploadCount(newUploadCount);
+    setIsLoading(true);
 
     if (userUUID) {
       localStorage.setItem(`uploadCount_${userUUID}`, newUploadCount.toString());
-    }
-
-    if (newUploadCount === 0) {
-      document.getElementById('submitBtn').disabled = true;
     }
 
     // Create a FormData object to send the file
@@ -91,19 +100,29 @@ const InputForm = () => {
     } catch (error) {
       // Handle errors here
       console.error('Error uploading file:', error.message);
+    } finally {
+      setIsLoading(false);
     }
   };
 
 
   return (
     <div className="input-form">
-      <br />
-      <input type="file" accept=".png, .jpg, .jpeg, .pdf" onChange={handleFileChange} />
-      <br />
-      <p>You have {uploadCount} submits left.</p>
-      <button id="submitBtn" onClick={handleUpload} disabled={!isFileSelected || uploadCount === 0}>
-        Submit
-      </button>
+      {isLoading ? (
+        <BorderExample />
+      ) : (
+        <>
+          <br />
+          <input type="file" accept=".png, .jpg, .jpeg, .pdf" onChange={handleFileChange} />
+          <br />
+          <br />
+          <br />
+          <p>You have {uploadCount} submits left.</p>
+          <button id="submitBtn" onClick={handleUpload} disabled={!isFileSelected || uploadCount === 0}>
+            Submit
+          </button>
+        </>
+      )}
 
       {/* Display the server response in the modal */}
       <ServerResponseModal
